# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal.MetaData:
    properties:
      current_page:
        type: integer
      first_page:
        type: integer
      last_page:
        type: integer
      page_size:
        type: integer
      total_records:
        type: integer
    type: object
  internal.Movie:
    properties:
      created_at:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      runtime:
        type: integer
      title:
        type: string
      version:
        type: integer
      year:
        type: integer
    type: object
...
  main.CreateMovieResponse:
    properties:
      movie:
        $ref: '#/definitions/internal.Movie'
    type: object
...
  main.GetMovieResponse:
    properties:
      movie:
        $ref: '#/definitions/internal.Movie'
    type: object
  main.GetMoviesResponse:
    properties:
      meta_data:
        $ref: '#/definitions/internal.MetaData'
      movies:
        items:
          $ref: '#/definitions/internal.Movie'
        type: array
    type: object
...
  main.ResponseError:
    properties:
      error:
        description: Error
        type: string
    type: object
  main.ResponseMessage:
    properties:
      message:
        description: Message
        type: string
    type: object
...
  main.UpdateMovieResponse:
    properties:
      movie:
        $ref: '#/definitions/internal.Movie'
    type: object
...
  main.ViolationsMessage:
    properties:
      errors:
        additionalProperties:
          type: string
        description: errors
        type: object
    type: object
host: https://localhost:8080
info:
  contact:
    email: ahamerdev@gmail.com
    name: Ahmed Amer
  description: a simple movie reservation system api for educational purposes
  title: Movie Reservation System API
  version: "1.0"
paths:
...
  /movies:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetMoviesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
    post:
      consumes:
      - application/json
      description: sreates a movie
      parameters:
      - description: title
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: runtime
        in: body
        name: runtime
        required: true
        schema:
          type: integer
      - description: year
        in: body
        name: year
        required: true
        schema:
          type: integer
      - description: genres
        in: body
        name: genres
        required: true
        schema:
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.CreateMovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Creates a movie
      tags:
      - movies
  /movies/{id}:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
    get:
      consumes:
      - application/json
      description: gets a movie by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.GetMovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
      summary: Gets a movie
      tags:
      - movies
    put:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UpdateMovieResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ViolationsMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/main.ResponseMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ResponseError'
...
swagger: "2.0"